CDH-53414: Sentry fails to setup secure connection to HMS if the service unix user is missing active tgt
Looking at the code of Sentry and HiveMetaStoreClient, it looks like
this problem has been there for a long time. But it's mostly masked by
the restarts we are doing after the clean deployment.

Some context on how our internal deployment works:
CDEP as part of deployment kicks off the first run of all services via
CM and once it is done, it kinit's service principals under the service
unix users for e.g: hdfs@VPC.CLOUDERA.COM gets activated under hdfs
user.

So during the course of CM first run, sentry and other services don't
have any keytabs active under local unix user and during this time, the
log of sentry is cluttered with the above errors. But after we do a
restart after CDEP completes, the Sentry code picks up the principal
that was activated under "sentry" user and thus works properly.

Ideally the service processes shouldn't depend on the keytabs active on
running unix user but rather use the keytabs supplied by CM in the
process directory.

Looking at the code of HMSFollower we invoke the classes of Hive -
HiveMetaStoreClient which exclusively uses the UserGroupInformation
object. And for UserGroupInformation object to pick up the keytab, one
should explicitly call the methods like loginUserFromKeytabAndReturnUGI
or use getUGIFromSubject. For the fact that HMSFollower already logs
into the keytab, we can leverage the method getUGIFromSubject() to make
the UserGroupInformation aware of keytab authentication.

If UGI object is not made aware of keytab, then the invocation of
UserGroupInformation#getCurrentUser falls back to Unix user.

Adding extra logic to HMSFollower to make UserGroupInformation aware
of the keytab that way it uses it for communication with HMS.

Change-Id: I520ef1183b55e286386d401ce6c1b3946a648d18
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/22294
Reviewed-by: Alexander Kolbasov <akolb@cloudera.com>
Tested-by: Alexander Kolbasov <akolb@cloudera.com>